Remove socket patching
